+++ RayTracerChallenge.Domain.Tests/MatrixComparerTests.cs
@@ -11,23 +11,23 @@ public class MatrixComparerTests : AbstractComparerTests<Matrix>
     [Fact]
     public void SameValuesAreEqual()
     {
-        var data = new float[4,4];
-        data[0,0] = 1.0F;
-        data[0,1] = 2.0F;
-        data[0,2] = 3.0F;
-        data[0,3] = 4.0F;
-        data[1,0] = 5.0F;
-        data[1,1] = 6.0F;
-        data[1,2] = 7.0F;
-        data[1,3] = 8.0F;
-        data[2,0] = 9.0F;
-        data[2,1] = 8.0F;
-        data[2,2] = 7.0F;
-        data[2,3] = 6.0F;
-        data[3,0] = 5.0F;
-        data[3,1] = 4.0F;
-        data[3,2] = 3.0F;
-        data[3,3] = 2.0F;
+        var data = new decimal[4,4];
+        data[0,0] = 1;
+        data[0,1] = 2;
+        data[0,2] = 3;
+        data[0,3] = 4;
+        data[1,0] = 5;
+        data[1,1] = 6;
+        data[1,2] = 7;
+        data[1,3] = 8;
+        data[2,0] = 9;
+        data[2,1] = 8;
+        data[2,2] = 7;
+        data[2,3] = 6;
+        data[3,0] = 5;
+        data[3,1] = 4;
+        data[3,2] = 3;
+        data[3,3] = 2;
 
         var matrix1 = new Matrix(data);
         var matrix2 = new Matrix(data);
